requirements.txt
openai
python-dotenv
pdfminer
pdfminer.six
bs4